+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All important changes to `fiisoft-jackdaw-stream` will be documented in this file
 
+## 2.1.0 - incompatible changes!
+
+- removed method ResultItem::create
+- added method StreamApi::collect
+- class FiiSoft\Jackdaw\Collector\Collectors\Collect renamed to CollectIn
+- composer.lock added to .gitignore
+
 ## 2.0.0
 
 Version for PHP 7.4, developed on branch php74